first upload
now only working with python for focus
"max_tokens": 10000, # default number of token from 200 to 10000
agent.sh to interact with it.
usage :
$ ./agent.sh input
where 'input' is the number of fire-typed pokemon you want to be served with in the JSON output.
Standard JSON ouput working with T=0.5 but not with all LLM models
removed other outputs but keeping them in code for futur debug mode or whatever
"temperature": 0.5, # default 0.0 to 0.5 for creativity lol
System prompt is now in LLM_SYSTEM_PROMPT environment variable
Default system prompt is : "Act as the backend of a Pokedex. Do exactly what you are told to do, a json file, and stricly that JSON file, without wording, notes or commentary. You can be as creative as you need to with categories."